Placed the available sites into an array
I limited the sites that a user could choose from by placing it into a
shared  array that both the ActionSheet would use as well as the logic
to determine the initial page load. Also added in a decisionHandler to
determine if a link could be followed off of one of the pages that a
user taps on by looking at its domain and seeing if the domain was on
the accepted list.